Move the server bits into its own folder.
