+++ packages/board-server/src/server/cors.ts
