+++ packages/board-server/src/server/errors.ts
